refactor: avatar badge
