+++ src/components/AvatarBadge/AvatarBadge.style.ts
@@ -1,43 +1,57 @@
-import { Avatar as MuiAvatar, Badge as MuiBadge } from '@mui/material';
+import { Avatar, Badge, Avatar as MuiAvatar } from '@mui/material';
 import { styled } from '@mui/material/styles';
-import { getAvatarMask } from '../Mask.style';
+import type { BadgeOffsetProps } from './AvatarBadge';
+import { getAvatarMask } from './getAvatarMask';
 
-// Styled component for Avatar with dynamic mask
-export const StyledAvatar = styled(MuiAvatar)<{
+interface StyledAvatarProps {
   avatarSize: number;
   badgeSize: number;
-}>(({ avatarSize, badgeSize }) => ({
-  margin: 'auto',
+  badgeOffset?: BadgeOffsetProps;
+  badgeGap?: number;
+}
+
+// Styled Avatar component for the badge
+export const StyledAvatar = styled(Avatar, {
+  shouldForwardProp: (prop) =>
+    prop !== 'avatarSize' &&
+    prop !== 'badgeSize' &&
+    prop !== 'badgeOffset' &&
+    prop !== 'badgeGap',
+})<StyledAvatarProps>(({ avatarSize, badgeSize, badgeOffset, badgeGap }) => ({
   height: avatarSize,
   width: avatarSize,
+  mask: getAvatarMask({ avatarSize, badgeSize, badgeOffset, badgeGap }), // Apply dynamic mask based on avatar and badge size
   '> img': {
     height: '100%',
     width: '100%',
     objectFit: 'contain',
   },
-  mask: getAvatarMask({ avatarSize, badgeSize }), // Applying the dynamic mask
 }));
 
-// Styled component for Badge Content (badge image)
-export const StyledBadgeContent = styled(MuiAvatar)<{ size: number }>(
-  ({ size }) => ({
-    width: size,
-    height: size,
-    backgroundColor: '#fff',
-    '& img': {
-      objectFit: 'contain',
-    },
-  }),
-);
+interface StyledBadgeProps {
+  badgeOffset?: BadgeOffsetProps;
+}
 
-// Styled component for AvatarBadge
-export const StyledAvatarBadge = styled(MuiBadge)<{
-  avatarSize: number;
-  badgeSize: number;
-}>(({ avatarSize, badgeSize }) => ({
+// Styled Badge component for the badge
+export const StyledBadge = styled(Badge, {
+  shouldForwardProp: (prop) => prop !== 'badgeOffset',
+})<StyledBadgeProps>(({ badgeOffset }) => ({
   borderRadius: '50%',
-  '> .MuiAvatar-root': {
-    mask: getAvatarMask({ avatarSize, badgeSize }), // Applying the same dynamic mask to the badge
+  display: 'block',
+
+  '.MuiBadge-badge': {
+    position: 'static',
+    transform: 'none',
+    top: 'unset',
+    right: 'unset',
+    zIndex: 'unset',
+    minWidth: 'unset',
+    padding: 'unset',
+    height: 'unset',
+    lineHeight: 'unset',
+    ...((badgeOffset?.x || badgeOffset?.y) && {
+      transform: `translate(${badgeOffset?.x ? badgeOffset.x : 0}px, ${badgeOffset?.y ? badgeOffset.y : 0}px)`,
+    }),
   },
 }));
 
@@ -52,5 +66,4 @@ export const StyledBadgeAvatar = styled(MuiAvatar)<{
   right: 0,
   top: 'unset',
   left: 'unset',
-  // display: 'block',
 }));
